chore(state): extract selectors and move items under slices
