+++ src/redux/slices/consent.ts
@@ -1,7 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
-import type { AppState } from '@/redux/store';
-
 export interface ConsentState {
   reviewCompleted: boolean;
   allowCookies: boolean;
@@ -47,16 +45,5 @@ export const {
   setAllowLocation,
   setAllowNotification,
 } = consent.actions;
-
-export const selectReviewCompleted = (state: AppState) =>
-  state.consent.reviewCompleted;
-export const selectAllowCookies = (state: AppState) =>
-  state.consent.allowCookies;
-export const selectAllowAnalytics = (state: AppState) =>
-  state.consent.allowAnalytics;
-export const selectAllowLocation = (state: AppState) =>
-  state.consent.allowLocation;
-export const selectAllowNotification = (state: AppState) =>
-  state.consent.allowNotification;
 
 export default consent.reducer;
